+++ FrskyGps/defines.h
@@ -1,12 +1,13 @@
 #include "Arduino.h"
 
-int        heading = 0;
-int        gps_alt = 0;
-int        groundspeed = 0;
+float      heading = 0;
+float      gps_alt = 0;
+float      groundspeed = 0.0f;
 float      lat = 0;
 float      lon = 0;
 uint8_t    sats = 0;
-int   fix = 0;
+uint16_t   fix = 0;
 
 byte lat_dir;
 byte lon_dir;
+
